move getMatrixDistances under Routing
