+++ src/index.js
@@ -4,7 +4,6 @@ import Navigator from './navigator';
 import Context from './api/context';
 import Geocoding from './api/geocoding';
 import Routing from './api/routing';
-import Matrix from './api/matrix';
 import Search from './api/search';
 import Track from './api/track';
 import Trips from './api/trips';
@@ -291,7 +290,7 @@ class Radar {
   }
 
   static getMatrix(routingOptions, callback=defaultCallback) {
-    Matrix.getMatrixDistances(routingOptions)
+    Routing.getMatrixDistances(routingOptions)
       .then((response) => {
         callback(null, { routes: response.routes, status: STATUS.SUCCESS }, response);
       })
